@@ -32,19 +32,17 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 import javax.faces.context.FacesContext;
 import javax.servlet.http.HttpServletRequest;
 
-import org.apache.commons.configuration2.ex.ConfigurationException;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.apache.solr.common.SolrDocument;
 import org.apache.solr.common.SolrDocumentList;
 import org.jdom2.JDOMException;
-import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.LogManager;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -312,11 +310,9 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
         switch (baseMimeType) {
             case IMAGE:
                 return getImageUrl();
-            case VIDEO:
-            case AUDIO: {
+            case VIDEO, AUDIO:
                 String format = getFileNames().keySet().stream().findFirst().orElse("");
                 return getMediaUrl(format);
-            }
             case APPLICATION:
                 if (StringUtils.isEmpty(fileName)) {
                     fileName = determineFileName(filePath);
@@ -370,14 +366,14 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
                             .search(new StringBuilder(SolrConstants.PI).append(":").append(pi).toString(), SolrSearchIndex.MAX_HITS, null,
                                     Collections.singletonList(field));
                     if (res != null && !res.isEmpty() && res.get(0).getFirstValue(field) != null) {
-                        // logger.debug(field + ":" + res.get(0).getFirstValue(field));
+                        // logger.debug("{}:{}", field, res.get(0).getFirstValue(field)); //NOSONAR Debug
                         urlBuilder.append((String) res.get(0).getFirstValue(field));
                         break;
                     }
                 } catch (PresentationException e) {
-                    logger.debug("PresentationException thrown here: " + e.getMessage());
+                    logger.debug("PresentationException thrown here: {}", e.getMessage());
                 } catch (IndexUnreachableException e) {
-                    logger.debug("IndexUnreachableException thrown here: " + e.getMessage());
+                    logger.debug("IndexUnreachableException thrown here: {}", e.getMessage());
 
                 }
             } else if (StringUtils.equalsIgnoreCase(text, WATERMARK_TEXT_TYPE_URN)) {
@@ -559,17 +555,16 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
 
     /**
      * 
-     * @return
+     * @return First part of the mime type
      * @should return correct base mime type
      * @should return image if base mime type not found
      */
     public String getBaseMimeType() {
         BaseMimeType baseMimeType = BaseMimeType.getByName(mimeType);
-        if(BaseMimeType.UNKNOWN.equals(baseMimeType)) {
+        if (BaseMimeType.UNKNOWN.equals(baseMimeType)) {
             return BaseMimeType.IMAGE.getName();
-        } else {
-            return baseMimeType.getName();            
         }
+        return baseMimeType.getName();
     }
 
     /**
@@ -640,7 +635,7 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
 
     /**
      *
-     * @return
+     * @return true if page has an indexed image file name and user has access permission; false otherwise
      * @throws DAOException
      * @throws IndexUnreachableException
      */
@@ -917,7 +912,7 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
                 fullText = ViewerResourceBundle.getTranslation(e.getMessage(), null);
             } catch (FileNotFoundException e) {
                 logger.error(e.getMessage());
-            } catch (IOException | IndexUnreachableException | DAOException e) {
+            } catch (IOException | IndexUnreachableException e) {
                 logger.error(e.getMessage(), e);
             }
         }
@@ -954,13 +949,11 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
      * @return true if fulltext loaded successfully false otherwise
      * @throws AccessDeniedException
      * @throws IOException
-     * @throws DAOException
      * @throws IndexUnreachableException
-     * @throws ConfigurationException
      * @should load full-text correctly if not yet loaded
      * @should return null if already loaded
      */
-    String loadFullText() throws AccessDeniedException, IOException, IndexUnreachableException, DAOException, ViewerConfigurationException {
+    String loadFullText() throws AccessDeniedException, IOException, IndexUnreachableException {
         if (fulltextFileName == null) {
             return null;
         } else if (Boolean.FALSE.equals(isFulltextAccessPermission())) {
@@ -993,10 +986,11 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
      * Returns word coordinates for words that start with any of the given search terms.
      *
      * @param searchTerms a {@link java.util.Set} object.
-     * @should load XML document if none yet set
+     * @param proximitySearchDistance
      * @param rotation a int.
      * @return a {@link java.util.List} object.
      * @throws io.goobi.viewer.exceptions.ViewerConfigurationException if any.
+     * @should load XML document if none yet set
      */
     public List<String> getWordCoords(Set<String> searchTerms, int proximitySearchDistance, int rotation) throws ViewerConfigurationException {
         if (searchTerms == null || searchTerms.isEmpty()) {
@@ -1057,9 +1051,7 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
                 wordCoordsFormat = CoordsFormat.ALTO;
             }
             return alto;
-        } catch (ContentNotFoundException e) {
-            logger.error(e.getMessage());
-        } catch (PresentationException e) {
+        } catch (ContentNotFoundException | PresentationException e) {
             logger.error(e.getMessage());
         }
 
@@ -1338,6 +1330,7 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
         BaseMimeType baseMimeType = BaseMimeType.getByName(this.mimeType);
         if (baseMimeType == null) {
             return "viewImage";
+
         }
 
         switch (baseMimeType) {
@@ -1364,7 +1357,7 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
     public boolean isAccessPermission3DObject() throws IndexUnreachableException, DAOException {
         logger.trace("AccessPermission3DObject");
         // Prevent access if mime type incompatible
-        if (!(BaseMimeType.OBJECT.equals(BaseMimeType.getByName(mimeType)) || BaseMimeType.MODEL.equals(BaseMimeType.getByName(mimeType)))) {
+        if (!BaseMimeType.MODEL.equals(BaseMimeType.getByName(mimeType))) {
             return false;
         }
 
@@ -1440,7 +1433,7 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
 
     /**
      *
-     * @return
+     * @return true if user has access permission; false otherwise
      * @throws IndexUnreachableException
      * @throws DAOException
      */
@@ -1497,10 +1490,10 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
             HttpServletRequest request = (HttpServletRequest) FacesContext.getCurrentInstance().getExternalContext().getRequest();
             AccessPermission access = AccessConditionUtils.checkAccessPermissionByIdentifierAndFileNameWithSessionMap(request, pi, fileName,
                     IPrivilegeHolder.PRIV_DOWNLOAD_BORN_DIGITAL_FILES);
-            // logger.trace("Born digital access for page {} is granted: {}", order, access.isGranted());
+            // logger.trace("Born digital access for page {} is granted: {}", order, access.isGranted()); //NOSONAR Debug
             //            if (bornDigitalDownloadTicketRequired == null) {
             bornDigitalDownloadTicketRequired = access.isTicketRequired();
-            // logger.trace("Ticket required for page {}: {}", order, access.isTicketRequired());
+            // logger.trace("Ticket required for page {}: {}", order, access.isTicketRequired()); //NOSONAR Debug
             //            }
             return access.isGranted();
         }
@@ -1520,13 +1513,13 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
         //                if (bornDigitalDownloadTicketRequired == null) {
         isAccessPermissionBornDigital();
         //                }
-        // logger.trace("isBornDigitalDownloadTicketRequired: {}", bornDigitalDownloadTicketRequired);
+        // logger.trace("isBornDigitalDownloadTicketRequired: {}", bornDigitalDownloadTicketRequired); //NOSONAR Debug
 
         // If license requires a download ticket, check agent session for loaded ticket
         if (Boolean.TRUE.equals(bornDigitalDownloadTicketRequired) && FacesContext.getCurrentInstance() != null
                 && FacesContext.getCurrentInstance().getExternalContext() != null) {
             boolean hasTicket = AccessConditionUtils.isHasDownloadTicket(pi, BeanUtils.getSession());
-            // logger.trace("User has download ticket: {}", hasTicket);
+            // logger.trace("User has download ticket: {}", hasTicket); //NOSONAR Debug
             return !hasTicket;
         }
 
@@ -1535,7 +1528,7 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
 
     /**
      *
-     * @return
+     * @return true if user has access permission; false otherwise
      * @throws IndexUnreachableException
      * @throws DAOException
      */
@@ -1631,7 +1624,7 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
     /**
      *
      * @param load If true, ALTO will be loaded if altoText is null
-     * @return
+     * @return ALTO document for this page
      * @throws ViewerConfigurationException
      */
     public String getAltoText(boolean load) throws ViewerConfigurationException {
@@ -1816,7 +1809,7 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
 
     /**
      *
-     * @return
+     * @return {@link String}
      * @throws PresentationException
      * @throws IndexUnreachableException
      * @throws JsonProcessingException
@@ -1829,7 +1822,7 @@ public class PhysicalElement implements Comparable<PhysicalElement>, Serializabl
             List<ShapeMetadata> shapes = elements.stream()
                     .filter(ele -> ele.getShapeMetadata() != null && !ele.getShapeMetadata().isEmpty())
                     .flatMap(ele -> ele.getShapeMetadata().stream())
-                    .collect(Collectors.toList());
+                    .toList();
             return mapper.writeValueAsString(shapes);
         }
     }
