@@ -783,6 +783,7 @@ public class ActiveDocumentBean implements Serializable {
      * @throws IndexUnreachableException
      */
     public void setRepresentativeImage() throws PresentationException, IndexUnreachableException {
+        logger.trace("setRepresentativeImage"); //NOSONAR Debug
         synchronized (lock) {
             String image = "1";
             if (StringUtils.isNotEmpty(lastReceivedIdentifier) && !"-".equals(lastReceivedIdentifier)) {
@@ -790,10 +791,10 @@ public class ActiveDocumentBean implements Serializable {
                         .getSearchIndex()
                         .getFirstDoc(SolrConstants.PI + ":" + lastReceivedIdentifier, Collections.singletonList(SolrConstants.THUMBPAGENO));
                 if (doc != null && doc.getFieldValue(SolrConstants.THUMBPAGENO) != null) {
-                    this.imageToShow = String.valueOf(doc.getFieldValue(SolrConstants.THUMBPAGENO));
-                    logger.trace("{} found: {}", SolrConstants.THUMBPAGENO, this.imageToShow);
+                    image = String.valueOf(doc.getFieldValue(SolrConstants.THUMBPAGENO));
+                    logger.trace("{} found: {}", SolrConstants.THUMBPAGENO, image);
                 } else {
-                    logger.trace("{}  not found, using {}", SolrConstants.THUMBPAGENO, this.imageToShow);
+                    logger.trace("{}  not found, using {}", SolrConstants.THUMBPAGENO, image);
                 }
             }
             setImageToShow(image);
@@ -1456,19 +1457,6 @@ public class ActiveDocumentBean implements Serializable {
         }
 
         return getPageUrl(PageType.viewFullscreen.getName(), imageToShow);
-    }
-
-    /**
-     * <p>
-     * getReadingModeUrl.
-     * </p>
-     *
-     * @deprecated renamed to fullscreen
-     * @return a {@link java.lang.String} object.
-     * @throws io.goobi.viewer.exceptions.IndexUnreachableException if any.
-     */
-    public String getReadingModeUrl() throws IndexUnreachableException {
-        return getFullscreenImageUrl();
     }
 
     /**
@@ -2730,5 +2718,4 @@ public class ActiveDocumentBean implements Serializable {
     public List<String> getGeomapFilters() {
         return List.of("MD_METADATATYPE", "MD_GENRE").stream().map(s -> "'" + s + "'").collect(Collectors.toList());
     }
-
 }
