@@ -30,12 +30,9 @@ import java.util.List;
 
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.io.FileUtils;
-import org.apache.solr.common.SolrDocument;
-import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.LogManager;
-
-import com.fasterxml.jackson.core.JsonParseException;
-import com.fasterxml.jackson.databind.JsonMappingException;
+import org.apache.logging.log4j.Logger;
+import org.apache.solr.common.SolrDocument;
 
 import de.intranda.api.annotation.wa.Motivation;
 import de.intranda.api.annotation.wa.WebAnnotation;
@@ -83,25 +80,26 @@ public class AnnotationIndexAugmenter implements IndexAugmenter {
     @Override
     public void augmentReIndexRecord(String pi, String dataRepository, String namingScheme) throws Exception {
 
-        Collection<PersistentAnnotation> annotations = this.annotations != null ? this.annotations : loadAllAnnotations(pi);
+        Collection<PersistentAnnotation> annos = this.annotations != null ? this.annotations : loadAllAnnotations(pi);
 
-        if (!annotations.isEmpty()) {
-            logger.debug("Found {} annotations for this record.", annotations.size());
-            writeToHotfolder(namingScheme, annotations);
+        if (!annos.isEmpty()) {
+            logger.debug("Found {} annotations for this record.", annos.size());
+            writeToHotfolder(namingScheme, annos);
         }
     }
 
     /* (non-Javadoc)
-     * @see io.goobi.viewer.modules.interfaces.IndexAugmenter#augmentReIndexPage(java.lang.String, int, org.apache.solr.common.SolrDocument, java.lang.String, java.lang.String)
+     * @see io.goobi.viewer.modules.interfaces.IndexAugmenter#augmentReIndexPage(java.lang.String, int, org.apache.solr.common.SolrDocument,
+     * java.lang.String, java.lang.String)
      */
     @Override
     public boolean augmentReIndexPage(String pi, int page, SolrDocument doc, String dataRepository, String namingScheme) throws Exception {
 
-        Collection<PersistentAnnotation> annotations = this.annotations != null ? this.annotations : loadAllAnnotations(pi, page);
+        Collection<PersistentAnnotation> annos = this.annotations != null ? this.annotations : loadAllAnnotations(pi, page);
 
-        if (!annotations.isEmpty()) {
-            logger.debug("Found {} annotations for this record.", annotations.size());
-            writeToHotfolder(namingScheme, annotations);
+        if (!annos.isEmpty()) {
+            logger.debug("Found {} annotations for this record.", annos.size());
+            writeToHotfolder(namingScheme, annos);
         }
         return true;
     }
@@ -114,31 +112,27 @@ public class AnnotationIndexAugmenter implements IndexAugmenter {
                 String json = webAnno.toString();
                 String jsonFileName = annotation.getTargetPI() + "_" + annotation.getId() + ".json";
                 FileUtils.writeStringToFile(new File(annotationDir, jsonFileName), json, Charset.forName(StringTools.DEFAULT_ENCODING));
-            } catch (JsonParseException e) {
-                logger.error(e.getMessage(), e);
-            } catch (JsonMappingException e) {
-                logger.error(e.getMessage(), e);
             } catch (IOException e) {
                 logger.error(e.getMessage(), e);
             }
         }
     }
 
-    private Collection<PersistentAnnotation> loadAllAnnotations(String pi) throws DAOException {
+    private static Collection<PersistentAnnotation> loadAllAnnotations(String pi) throws DAOException {
         List<PersistentAnnotation> annotations = new ArrayList<>();
         annotations.addAll(loadAllCampaignAnnotations(pi));
         annotations.addAll(loadAllCommentAnnotations(pi));
         return annotations;
     }
 
-    private Collection<PersistentAnnotation> loadAllAnnotations(String pi, int page) throws DAOException {
+    private static Collection<PersistentAnnotation> loadAllAnnotations(String pi, int page) throws DAOException {
         List<PersistentAnnotation> annotations = new ArrayList<>();
         annotations.addAll(loadAllCampaignAnnotations(pi, page));
         annotations.addAll(loadAllCommentAnnotations(pi, page));
         return annotations;
     }
 
-    private Collection<PersistentAnnotation> loadAllCampaignAnnotations(String pi) throws DAOException {
+    private static Collection<PersistentAnnotation> loadAllCampaignAnnotations(String pi) throws DAOException {
         List<CampaignRecordStatistic> statistics =
                 DataManager.getInstance().getDao().getCampaignStatisticsForRecord(pi, CrowdsourcingStatus.FINISHED);
         List<PersistentAnnotation> annotations = new ArrayList<>();
@@ -158,7 +152,7 @@ public class AnnotationIndexAugmenter implements IndexAugmenter {
         return annotations;
     }
 
-    private Collection<CrowdsourcingAnnotation> loadAllCampaignAnnotations(String pi, int page) throws DAOException {
+    private static Collection<CrowdsourcingAnnotation> loadAllCampaignAnnotations(String pi, int page) throws DAOException {
         List<CampaignRecordPageStatistic> pageStatistics =
                 DataManager.getInstance().getDao().getCampaignPageStatisticsForRecord(pi, CrowdsourcingStatus.FINISHED);
         CampaignRecordPageStatistic statistic = pageStatistics.stream()
@@ -166,18 +160,15 @@ public class AnnotationIndexAugmenter implements IndexAugmenter {
                 .findAny()
                 .orElseThrow(() -> new IllegalStateException("No page number " + page + " found"));
         Campaign campaign = statistic.getOwner().getOwner();
-        List<CrowdsourcingAnnotation> annotations = DataManager.getInstance().getDao().getAnnotationsForCampaignAndTarget(campaign, pi, page);
-        return annotations;
+        return DataManager.getInstance().getDao().getAnnotationsForCampaignAndTarget(campaign, pi, page);
     }
 
-    private Collection<Comment> loadAllCommentAnnotations(String pi, int page) throws DAOException {
-        List<Comment> comments = DataManager.getInstance().getDao().getCommentsForPage(pi, page);
-        return comments;
+    private static Collection<Comment> loadAllCommentAnnotations(String pi, int page) throws DAOException {
+        return DataManager.getInstance().getDao().getCommentsForPage(pi, page);
     }
 
-    private Collection<CrowdsourcingAnnotation> loadAllCommentAnnotations(String pi) throws DAOException {
-        List<CrowdsourcingAnnotation> annos = DataManager.getInstance().getDao().getAnnotationsForTarget(pi, null, Motivation.COMMENTING);
-        return annos;
+    private static Collection<CrowdsourcingAnnotation> loadAllCommentAnnotations(String pi) throws DAOException {
+        return DataManager.getInstance().getDao().getAnnotationsForTarget(pi, null, Motivation.COMMENTING);
     }
 
     @Override
@@ -192,9 +183,9 @@ public class AnnotationIndexAugmenter implements IndexAugmenter {
     public boolean equals(Object obj) {
         if (obj != null && obj.getClass().equals(AnnotationIndexAugmenter.class)) {
             return CollectionUtils.isEqualCollection(this.annotations, ((AnnotationIndexAugmenter) obj).annotations);
-        } else {
-            return false;
         }
+
+        return false;
     }
 
 }
