@@ -28,6 +28,7 @@ import java.io.IOException;
 
 import javax.servlet.http.HttpServletRequest;
 
+import de.unigoettingen.sub.commons.contentlib.imagelib.ImageFileFormat;
 import de.unigoettingen.sub.commons.contentlib.imagelib.transform.Region;
 import io.goobi.viewer.api.rest.AbstractApiUrlManager;
 import io.goobi.viewer.api.rest.v1.authentication.UserAvatarResource;
@@ -66,8 +67,14 @@ public class LocalUserAvatar implements UserAvatar {
         try {
             String sizeString = "!" + size + "," + size;
             String format = UserAvatarResource.getAvatarFileSuffix(userId);
+            String displayFormat = format;
+            ImageFileFormat fileFormat = ImageFileFormat.getImageFileFormatFromFileExtension(format);
+            if(fileFormat != null) {
+                displayFormat = ImageFileFormat.getMatchingTargetFormat(fileFormat).getFileExtension();                
+            }
+            
             return urls.path(USERS_USER_AVATAR_IMAGE, USERS_USER_AVATAR_IMAGE_IIIF)
-                    .params(userId, Region.FULL_IMAGE, sizeString, 0, "default", format)
+                    .params(userId, Region.FULL_IMAGE, sizeString, 0, "default", displayFormat)
                     .query("updated", this.updated)
                     //                    .query("timestamp", System.currentTimeMillis())
                     .build();
